requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
sqlalchemy[asyncio]==2.0.35 
asyncpg==0.29.0 
pyjwt==2.9.0 
passlib[bcrypt]==1.7.4 
pydantic-settings==2.5.2
pydantic[email]
asyncpg